Fixes for review feedback

+++ discovery/src/main/java/io/airlift/discovery/client/CachingServiceSelector.java
@@ -157,7 +157,7 @@ public class CachingServiceSelector
             {
                 refresh();
             }
-        }, (long) delay.toMillis(), TimeUnit.MILLISECONDS);
+        }, delay.toMillis(), TimeUnit.MILLISECONDS);
     }
 
     private static <V> ListenableFuture<V> chainedCallback(
