+++ sample-server/src/main/java/io/airlift/sample/PersonStore.java
@@ -40,7 +40,7 @@ public class PersonStore
         Preconditions.checkNotNull(eventClient, "eventClient is null");
 
         Cache<String, Person> personCache = CacheBuilder.newBuilder()
-                .expireAfterWrite((long) config.getTtl().toMillis(), TimeUnit.MILLISECONDS)
+                .expireAfterWrite(config.getTtl().toMillis(), TimeUnit.MILLISECONDS)
                 .build();
         persons = personCache.asMap();
         stats = new PersonStoreStats(eventClient);
